No message together with link when sending from ActionBar.
Sending a message followed by the link works fine for all
intent-receivers except Bluetooth. In the case of Bluetooth
will the following happen. The message will be converted
into an html-file. The data that is provided will be used
as both link-name and link-target, thus generating an
unusable html-page. Therefore, has I removed the messagepart.
More info about BT code can be found in the last function of:
http://hi-android.info/src/com/android/bluetooth/opp/BluetoothOppLauncherActivity.java.html